requirements.txt
alabaster==0.7.16
asgiref==3.8.1
babel==2.16.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
Django==5.0.7
django-cors-headers==4.4.0
django-nose==1.4.7
djangorestframework==3.15.2
docutils==0.20.1
idna==3.8
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==2.1.5
nose==1.3.7
packaging==24.1
Pygments==2.18.0
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
